feat: 0으로 나누면 오류를 뱉을 수 있도록 구현 #5 - CalculationError 파일 생성 및 오류 생성 - 기존의 테스트 코드에 try!를 붙여 오류가 안나는 계산 테스트 - 0으로 나눴을 때 오류가 잘 나오는지 확인
