+++ Calculator/Calculator/Model/CalculationError.swift
@@ -1,0 +1,12 @@
+//
+//  CalculationError.swift
+//  Calculator
+//
+//  Created by 양호준 on 2021/11/14.
+//
+
+import Foundation
+
+enum CalculationError: Error {
+    case wrongCalculation
+}
