+++ Calculator/Calculator/Model/Operator.swift
@@ -1,0 +1,48 @@
+//
+//  Operator.swift
+//  Calculator
+//
+//  Created by 양호준 on 2021/11/14.
+//
+
+import Foundation
+
+enum Operator: Character, CaseIterable, CalculateItem {
+    case add = "➕"
+    case subtract = "➖"
+    case divide = "➗"
+    case multiply = "✖️"
+    
+    func calculate(lhs: Double, rhs: Double) throws -> Double {
+        switch self {
+        case .add:
+            return add(lhs: lhs, rhs: rhs)
+        case .subtract:
+            return subtract(lhs: lhs, rhs: rhs)
+        case .divide:
+            if rhs == 0 {
+                throw CalculationError.wrongCalculation
+            } else {
+                return divide(lhs: lhs, rhs: rhs)
+            }
+        case .multiply:
+            return multiply(lhs: lhs, rhs: rhs)
+        }
+    }
+    
+    private func add(lhs: Double, rhs: Double) -> Double {
+        return lhs + rhs
+    }
+    
+    private func subtract(lhs: Double, rhs: Double) -> Double {
+        return lhs - rhs
+    }
+    
+    private func divide(lhs: Double, rhs: Double) -> Double {
+        return lhs / rhs
+    }
+    
+    private func multiply(lhs: Double, rhs: Double) -> Double {
+        return lhs * rhs
+    }
+}
