+++ Calculator/Calculator/Model/QueueError.swift
@@ -9,5 +9,4 @@ import Foundation
 
 enum QueueError: Error {
     case emptyItem
-    case wrongCalculation
 }
